genealogy-research-tool/backend/requirements.txt
# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# LLM
openai==1.3.7

# Fuzzy matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
metaphone==0.6

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
httpx==0.25.2
